Cleanup and Optimizations
